// src/models/shift-start-interface.ts
export interface ShiftStartdatainterface {
    shift_start_date_time:string;
    lineId: string;
    stationId: string;
    equipmentGroupId: number;
    equepmentId:number;
    acquirerId: number;
    operatorId: number;
    terminalId: string;
    agentId   :number;
    tom_Efo_Shift_start:string;
    
  }